Don't treat numbers as lists.
Sometimes a document will contain text that looks like this:

```
Software version:

20230101.
```

Currently cmark-gfm will output the following:

```
<ol start=20230101>
<li></li>
</ol>
```

This isn't what the author had intended. The author intended for the
markdown to render as:

```
<p>20230101.</p>
```

If the blank item would be added to an existing list it is preserved.
For example:

```
1. Hello
2.
3. World
```

Renders as:

```
<ol>
<li>Hello</li>
<li></li>
<li>World</li>
</ol>
```